pyproject.toml
[tool.poetry]
name = "blurred-horse-bot"
version = "0.1.0"
description = ""
authors = ["Astrid Yu <astrid@astrid.tech>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "blurred_horse_bot"}]

[tool.poetry.scripts]
blurred-horse-bot = "blurred_horse_bot:cli"

[tool.poetry.dependencies]
python = "^3.10"
flickr-api = "^0.7.6"
pillow = "^9.4.0"
click = "^8.1.3"
mastodon-py = "^1.8.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^65.7.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
